m v1

+++ src/main/java/com/del/ministry/view/forms/DistrictListIForm.java
@@ -1,6 +1,7 @@
 package com.del.ministry.view.forms;
 
 import com.del.ministry.dao.ServiceManager;
+import com.del.ministry.db.AddressType;
 import com.del.ministry.db.District;
 import com.del.ministry.db.DistrictAddress;
 import com.del.ministry.utils.CommonException;
@@ -8,6 +9,7 @@ import com.del.ministry.utils.StringUtil;
 import com.del.ministry.view.Launcher;
 import com.del.ministry.view.actions.ObservableIPanel;
 import com.del.ministry.view.actions.ShowIFrameActionListener;
+import com.del.ministry.view.models.AddressTypeItem;
 import com.del.ministry.view.models.DistrictAddressItem;
 import com.del.ministry.view.models.DistrictNumbers;
 import com.del.ministry.view.models.SelectItemsModel;
@@ -18,6 +20,8 @@ import com.jgoodies.forms.factories.Paddings;
 
 import javax.swing.*;
 import java.awt.*;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
 import java.util.List;
 import java.util.*;
 import java.util.stream.Collectors;
@@ -32,6 +36,7 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
     private JButton newBtn;
     private JButton changeBtn;
     private JButton deleteBtn;
+    private JButton editBtn;
     private JButton deleteAddressBtn;
     private JButton generateBtn;
 
@@ -45,10 +50,20 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
         districtNumbersF = new JComboBox<>();
         districtAddressF = new JList<>();
         districtAddressF.setSelectionMode(ListSelectionModel.MULTIPLE_INTERVAL_SELECTION);
+        districtAddressF.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent evt) {
+                if (evt.getClickCount() == 2) {
+                    editEvent(districtAddressF.getSelectedValue().getAddress());
+                }
+            }
+        });
         newDistrictNumberF = new JTextField();
 
         addBtn = new JButton("Добавить");
         deleteBtn = new JButton("Удалить");
+        editBtn = new JButton("Редактировать");
+        editBtn.setEnabled(false);
         deleteAddressBtn = new JButton("Удалить");
         generateBtn = new JButton("Заполнить");
         newBtn = new JButton("Новый");
@@ -60,17 +75,18 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
         deleteAddressBtn.setEnabled(false);
 
         add(FormBuilder.create().
-                        columns("50, 5, 80, 5, 50, pref:grow, 100, 150, 5, 100")
-                        .rows("p, 5, p, 5, p, 5, p, 5, p, fill:100:grow, 5, p, p")
+                        columns("50, 5, 80, 5, 50, pref:grow, 100, 150, 5, 150")
+                        .rows("p, 5, p, 5, p, 5, p, 5, p, 5, p, fill:100:grow, 5, p, p")
                         .padding(Paddings.DIALOG)
                         .add("Участок:").xy(1, 1).add(districtNumbersF).xy(3, 1).add(districtSize).xy(5, 1).add(deleteBtn).xy(10, 1)
                         .addSeparator("Адреса").xyw(1, 3, 10)
-                        .add(districtAddressF).xywh(1, 5, 8, 6)
+                        .add(districtAddressF).xywh(1, 5, 8, 8)
                         .add(addBtn).xy(10, 5)
                         .add(deleteAddressBtn).xy(10, 7)
-                        .add(generateBtn).xy(10, 9)
-                        .addSeparator("Редактор").xyw(1, 12, 10)
-                        .add(newDistrictNumberF).xyw(1, 13, 3).add(changeBtn).xy(8, 13).add(newBtn).xy(10, 13)
+                        .add(editBtn).xy(10, 9)
+                        .add(generateBtn).xy(10, 11)
+                        .addSeparator("Редактор").xyw(1, 14, 10)
+                        .add(newDistrictNumberF).xyw(1, 15, 3).add(changeBtn).xy(8, 15).add(newBtn).xy(10, 15)
                         .build(),
                 BorderLayout.CENTER);
 
@@ -89,9 +105,14 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
             instance.setDistrict(selectedDistrict.getDistrict());
         });
 
+        editBtn.addActionListener(e -> editEvent(districtAddressF.getSelectedValue().getAddress()));
+
         addBtn.addActionListener(addAddressFormAction);
         generateBtn.addActionListener(generateFormAction);
-        districtAddressF.addListSelectionListener(e -> deleteAddressBtn.setEnabled(districtAddressF.getSelectedIndex() > -1));
+        districtAddressF.addListSelectionListener(e -> {
+            deleteAddressBtn.setEnabled(districtAddressF.getSelectedIndex() > -1);
+            editBtn.setEnabled(districtAddressF.getSelectedIndex() > -1);
+        });
         districtNumbersF.addActionListener(e -> initAddressList());
         deleteAddressBtn.addActionListener(e -> deleteAddress());
         deleteBtn.addActionListener(e -> deleteDistrict());
@@ -106,22 +127,32 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
         initAddressList();
     }
 
+    @Override
+    public String getTitle() {
+        return "Редактирование участков";
+    }
+
     private void renameDistrict() {
         if (districtNumbersF.getSelectedIndex() > -1) {
             DistrictNumbers selectedItem = districtNumbersF.getItemAt(districtNumbersF.getSelectedIndex());
             String number = newDistrictNumberF.getText();
             if (!StringUtil.isTrimmedEmpty(number)) {
                 District district = selectedItem.getDistrict();
-                district.setNumber(number);
-                try {
-                    ServiceManager.getInstance().updateDistrict(district);
-                    Launcher.mainFrame.setStatusText("Участок успешно переименован");
-                    Launcher.mainFrame.initLeftSideTree();
-                } catch (Exception e) {
-                    Launcher.mainFrame.setStatusError("Невозможно переименовать участок!", e);
-                } finally {
-                    initDistrictList(district.getId());
-                    newDistrictNumberF.setText("");
+                int answer = JOptionPane.showConfirmDialog(Launcher.mainFrame,
+                        String.format("Переименовать участок %s на %s?", district.getNumber(), number),
+                        "Удаление", JOptionPane.YES_NO_OPTION);
+                if (answer == JOptionPane.YES_OPTION) {
+                    district.setNumber(number);
+                    try {
+                        ServiceManager.getInstance().updateDistrict(district);
+                        Launcher.mainFrame.setStatusText("Участок успешно переименован");
+                        Launcher.mainFrame.initLeftSideTree();
+                    } catch (Exception e) {
+                        Launcher.mainFrame.setStatusError("Невозможно переименовать участок!", e);
+                    } finally {
+                        initDistrictList(district.getId());
+                        newDistrictNumberF.setText("");
+                    }
                 }
             }
         }
@@ -130,16 +161,21 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
     private void deleteDistrict() {
         DistrictNumbers selectedItem = districtNumbersF.getItemAt(districtNumbersF.getSelectedIndex());
         if (selectedItem != null) {
+            String number = selectedItem.getDistrict().getNumber();
             try {
-                ServiceManager.getInstance().deleteDistrict(selectedItem.getDistrict().getId());
-                Launcher.mainFrame.setStatusText("Запись удалена");
-                initDistrictList();
-                initAddressList();
-                Launcher.mainFrame.initLeftSideTree();
+                int answer = JOptionPane.showConfirmDialog(Launcher.mainFrame, String.format("Удалить участок %s?", number),
+                        "Удаление", JOptionPane.YES_NO_OPTION);
+                if (answer == JOptionPane.YES_OPTION) {
+                    ServiceManager.getInstance().deleteDistrict(selectedItem.getDistrict().getId());
+                    Launcher.mainFrame.setStatusText(String.format("Участок %s был удален", number));
+                    initDistrictList();
+                    initAddressList();
+                    Launcher.mainFrame.initLeftSideTree();
+                }
             } catch (CommonException e1) {
                 Launcher.mainFrame.setStatusError(e1.getMessage(), e1);
             } catch (Exception e1) {
-                Launcher.mainFrame.setStatusError("Ошибка при удалении участка", e1);
+                Launcher.mainFrame.setStatusError(String.format("Ошибка при удалении участка %s", number), e1);
             }
         }
     }
@@ -153,7 +189,7 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
                 ServiceManager.getInstance().createDistrict(d);
                 Launcher.mainFrame.setStatusText("Участок создан");
                 newDistrictNumberF.setText("");
-                initDistrictList();
+                initDistrictList(d.getId());
                 Launcher.mainFrame.initLeftSideTree();
             } catch (Exception e1) {
                 Launcher.mainFrame.setStatusError("Ошибка при сохранении участка", e1);
@@ -199,7 +235,7 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
                 List<DistrictAddress> addresses = ServiceManager.getInstance().findDistrictAddresses(district.getId());
                 List<DistrictAddressItem> collect = addresses.stream().
                         map(DistrictAddressItem::new).sorted().collect(Collectors.toList());
-                districtAddressF.setModel(new LinkedListModel<DistrictAddressItem>(collect));
+                districtAddressF.setModel(new LinkedListModel<>(collect));
                 printSize(collect.size());
             } catch (Exception e) {
                 Launcher.mainFrame.setStatusError("Нет доступа к адресам участка", e);
@@ -214,6 +250,55 @@ public class DistrictListIForm extends ObservableIPanel implements Observer {
     @Override
     public void update(Observable o, Object arg) {
         initAddressList();
+    }
+
+    private void editEvent(DistrictAddress da) {
+        JDialog i = new JDialog(Launcher.mainFrame, "Редактировать тип адреса", Dialog.ModalityType.APPLICATION_MODAL);
+        try {
+            JButton cancelBtn = new JButton("Отмена");
+            JButton saveBtn = new JButton("Сохранить");
+            JComboBox<AddressTypeItem> typeF = new JComboBox<>();
+            List<AddressType> addressTypes = ServiceManager.getInstance().findAddressTypes();
+            List<AddressTypeItem> types = addressTypes.stream().map(AddressTypeItem::new).collect(Collectors.toList());
+            AddressTypeItem selectedType = types.stream().filter(p -> p.getAddressType().getId().equals(da.getType().getId())).findFirst().orElse(null);
+            DefaultComboBoxModel<AddressTypeItem> model = new DefaultComboBoxModel<>(new Vector<>(types));
+            model.setSelectedItem(selectedType);
+            typeF.setModel(model);
+            JPanel buttons = new JPanel();
+            cancelBtn.addActionListener(e1 -> i.dispose());
+            saveBtn.addActionListener(e1 -> {
+                AddressTypeItem si = (AddressTypeItem) typeF.getSelectedItem();
+                if (si != null) {
+                    da.setType(si.getAddressType());
+                    try {
+                        ServiceManager.getInstance().updateDistrictAddress(da);
+                        initAddressList();
+                        i.dispose();
+                    } catch (CommonException e) {
+                        Launcher.mainFrame.setStatusError("Ошибка сохранения типа адреса", e);
+                    }
+                }
+            });
+            i.getContentPane().add(FormBuilder.create().
+                    columns("100, 5, 200")
+                    .rows("p, 5, p")
+                    .padding(Paddings.DIALOG)
+                    .addLabel("Тип адреса").xy(1, 1).add(typeF).xy(3, 1)
+                    .build(), BorderLayout.CENTER);
+            i.getContentPane().add(buttons, BorderLayout.SOUTH);
+
+            buttons.add(saveBtn);
+            buttons.add(cancelBtn);
+
+            i.setDefaultCloseOperation(WindowConstants.HIDE_ON_CLOSE);
+            i.pack();
+            i.setLocationRelativeTo(Launcher.mainFrame);
+            i.setVisible(true);
+
+        } catch (CommonException e1) {
+            Launcher.mainFrame.setStatusError("Ошибка редактирования типа адреса", e1);
+        }
+
     }
 
 }
